requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
Faker==8.11.0
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0